# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Basic RESTful API with MongoDB",
      "version": "1.0.0",
      "description": "A simple REST API using Node.js, Express, MongoDB, and Swagger"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/api/users": {
        "get": {
          "summary": "Retrieve a list of users",
          "responses": {
            "200": {
              "description": "A list of users",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new user",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "The created user"
            },
            "400": {
              "description": "Validation error"
            }
          }
        }
      },
      "/api/users/{id}": {
        "get": {
          "summary": "Get a user by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The user ID"
            }
          ],
          "responses": {
            "200": {
              "description": "A single user",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "404": {
              "description": "User not found"
            }
          }
        },
        "put": {
          "summary": "Update a user by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The user ID"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "The updated user"
            },
            "400": {
              "description": "Validation error"
            },
            "404": {
              "description": "User not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a user by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "The user ID"
            }
          ],
          "responses": {
            "200": {
              "description": "The deleted user"
            },
            "404": {
              "description": "User not found"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "required": [
            "name",
            "email"
          ],
          "properties": {
            "id": {
              "type": "string",
              "description": "The auto-generated ID of the user"
            },
            "name": {
              "type": "string",
              "description": "The name of the user"
            },
            "email": {
              "type": "string",
              "description": "The email of the user"
            }
          },
          "example": {
            "id": "60d0fe4f5311236168a109ca",
            "name": "John Doe",
            "email": "john@example.com"
          }
        }
      }
    }
  }
  